feat(speech): Add simple text to speech method.
Using OpenAI's whisper model, we can now convert speech to text.

Method needs to be updated so that it uses a 'on' and 'off' feature. Currently records audio, waits for 2 seconds of silence then processes it.

Known bug: silence is always recorded in 2 second batches. Silence shouldn't be processed at all.